Move the cache package out of the Mixer directory to make it more broadly available.
