+++ mixer/adapter/servicecontrol/BUILD
@@ -21,10 +21,10 @@ go_library(
         "//mixer/adapter/servicecontrol/config:go_default_library",
         "//mixer/adapter/servicecontrol/template/servicecontrolreport:go_default_library",
         "//mixer/pkg/adapter:go_default_library",
-        "//mixer/pkg/cache:go_default_library",
         "//mixer/pkg/status:go_default_library",
         "//mixer/template/apikey:go_default_library",
         "//mixer/template/quota:go_default_library",
+        "//pkg/cache:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
         "@com_github_pborman_uuid//:go_default_library",
