+++ mixer/adapter/servicecontrol/checkprocessor.go
@@ -27,9 +27,9 @@ import (
 	rpc "istio.io/gogo-genproto/googleapis/google/rpc"
 	"istio.io/istio/mixer/adapter/servicecontrol/config"
 	"istio.io/istio/mixer/pkg/adapter"
-	"istio.io/istio/mixer/pkg/cache"
 	"istio.io/istio/mixer/pkg/status"
 	"istio.io/istio/mixer/template/apikey"
+	"istio.io/istio/pkg/cache"
 )
 
 // checkImpl implements checkProcessor interface, handles doCheck call to Google ServiceControl backend.
