+++ mixer/adapter/servicecontrol/servicecontrol.go
@@ -27,9 +27,9 @@ import (
 	"istio.io/istio/mixer/adapter/servicecontrol/config"
 	"istio.io/istio/mixer/adapter/servicecontrol/template/servicecontrolreport"
 	"istio.io/istio/mixer/pkg/adapter"
-	"istio.io/istio/mixer/pkg/cache"
 	"istio.io/istio/mixer/template/apikey"
 	"istio.io/istio/mixer/template/quota"
+	"istio.io/istio/pkg/cache"
 )
 
 // servicecontrol adapter builder
